add bignum scaffolding
